Create 2d-range-search.py
The problem of 2D range search is basically to find an efficient method to search for the number of points inside a 2D range(rectangular box). The algorithm is quite straight forward. We repeatedly select the median of points based on their 'x' and 'y' coordinates alternatively and divide all the points based on that. We consecutively keep building this grid until we have one point max inside each cell of the grid. Then we check which of these cells of the grid are covered by the range query(rectangular box), and output the points accordingly. 

Basic Requirements for running this python file:
1. Python 3
2. Matplotlib Library