Create 1D-range-search.py
Essentially this is just a program which gives a range of numbers as an output depending on your query from an array. The naïve approach to this problem is quite straightforward. But when we use a BST(Binary Search Tree) to solve this problem we increase its efficiency exponentially. 

 Requirements:
Python 3